css/css-overscroll-behavior/parsing/overscroll-behavior-y-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Overscroll Behavior: parsing overscroll-behavior-y with valid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-overscroll-behavior-y/#propdef-overscroll-behavior-y">
<meta name="assert" content="overscroll-behavior-y supports the full grammar 'contain | none | auto'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("overscroll-behavior-y", "contain");
test_valid_value("overscroll-behavior-y", "none");
test_valid_value("overscroll-behavior-y", "auto");
</script>
</body>
</html>
